Add a main class to reproduce the bug
Bug:

In finch 0.9.2 there appears to be a bug when passing a request
with a Request body created with Buf.Utf8. The main class shows
that the bug isn't present when using the raw finagle-http API.
It also isn't present in finch 0.9.1 or 0.9.0, which is demonstrated
by uncommenting the `import io.finch.request._` and changing the
versions build.sbt.